Split listing card into own component
